Move function print_reports from __main__.py to writer/__init__.py.
